[add] init.go (#21)
